refactor(api): share the state vocabulary and the brief across backends
The neutral AgentTaskState lived in GitHub's client, and the managed backend
imported buildPrompt from copilot-backend.ts — so the second backend looked
like it depended on the first. Both move to neutral modules.

The prompt now takes a DeliveryContract, because the managed backend was
telling pulled rounds to upload over the build channel while reading their
output directory. A pulled round is told where to write, and is told plainly
that it has no progress channel, no inbox and no gate verdict.

Comment prose from the moved code becomes docs/build-brief.md; the touched
files ratchet their baselines down.
