+++ apps/api/src/managed-agent.ts
@@ -1,5 +1,5 @@
 // Vendor-neutral seam for hosted coding-agent platforms.
-import type { AgentTaskState } from './agent-tasks.js';
+import type { AgentTaskState } from './agent-state.js';
 
 // Coarse reasoning budget; vendors name it differently.
 export type ManagedAgentEffort = 'low' | 'medium' | 'high';
