+++ eslint-rules/comment-prose-baseline.json
@@ -41,7 +41,7 @@
     "apps/api/src/agent-game-key.ts": 346,
     "apps/api/src/agent-session-revocation.ts": 81,
     "apps/api/src/agent-tasks.test.ts": 197,
-    "apps/api/src/agent-tasks.ts": 1183,
+    "apps/api/src/agent-tasks.ts": 1135,
     "apps/api/src/agent-token.test.ts": 16,
     "apps/api/src/agent-token.ts": 638,
     "apps/api/src/app.test.ts": 386,
@@ -61,7 +61,7 @@
     "apps/api/src/bounded-map.test.ts": 28,
     "apps/api/src/bounded-map.ts": 208,
     "apps/api/src/builder.test.ts": 27,
-    "apps/api/src/builder.ts": 442,
+    "apps/api/src/builder.ts": 239,
     "apps/api/src/canonical-app-url.ts": 113,
     "apps/api/src/canonical-base64.ts": 56,
     "apps/api/src/canonical-host.test.ts": 54,
@@ -71,7 +71,7 @@
     "apps/api/src/code-lane.ts": 2069,
     "apps/api/src/contact.ts": 238,
     "apps/api/src/copilot-backend.test.ts": 510,
-    "apps/api/src/copilot-backend.ts": 1251,
+    "apps/api/src/copilot-backend.ts": 976,
     "apps/api/src/creation-limits.test.ts": 208,
     "apps/api/src/creation-limits.ts": 1037,
     "apps/api/src/creator-agent-key-routes.test.ts": 635,
@@ -166,13 +166,13 @@
     "apps/api/src/mcp-debug-log.ts": 110,
     "apps/api/src/mcp-install-links.ts": 214,
     "apps/api/src/mcp-oauth-metadata.test.ts": 73,
-    "apps/api/src/mcp-oauth-metadata.ts": 331,
+    "apps/api/src/mcp-oauth-metadata.ts": 319,
     "apps/api/src/mcp-open-round.test.ts": 71,
     "apps/api/src/mcp-presence.ts": 224,
     "apps/api/src/mcp-server-discovery.test.ts": 90,
     "apps/api/src/mcp-server-discovery.ts": 316,
-    "apps/api/src/mcp-server.test.ts": 1540,
-    "apps/api/src/mcp-server.ts": 4564,
+    "apps/api/src/mcp-server.test.ts": 1533,
+    "apps/api/src/mcp-server.ts": 4535,
     "apps/api/src/mcp-session-key.ts": 287,
     "apps/api/src/mcp-session-nudges.test.ts": 17,
     "apps/api/src/mcp-session-nudges.ts": 494,
@@ -212,7 +212,7 @@
     "apps/api/src/owner-of-record.ts": 429,
     "apps/api/src/player-feedback.test.ts": 72,
     "apps/api/src/player-feedback.ts": 646,
-    "apps/api/src/plugin-manifests.test.ts": 354,
+    "apps/api/src/plugin-manifests.test.ts": 289,
     "apps/api/src/presence.test.ts": 457,
     "apps/api/src/presence.ts": 1597,
     "apps/api/src/proposal-apply-bot.test.ts": 34,
@@ -255,11 +255,11 @@
     "apps/api/src/seed-context.ts": 585,
     "apps/api/src/seed-metrics.test.ts": 311,
     "apps/api/src/seed-metrics.ts": 358,
-    "apps/api/src/seed-status.ts": 64,
+    "apps/api/src/seed-status.ts": 58,
     "apps/api/src/self-build-backend.ts": 263,
     "apps/api/src/self-build-connect.test.ts": 45,
     "apps/api/src/self-build-connect.ts": 507,
-    "apps/api/src/self-build.test.ts": 316,
+    "apps/api/src/self-build.test.ts": 315,
     "apps/api/src/server.ts": 110,
     "apps/api/src/slug-backfill.test.ts": 81,
     "apps/api/src/slug-backfill.ts": 335,
@@ -278,7 +278,7 @@
     "apps/api/src/submission-status.test.ts": 35,
     "apps/api/src/submission-status.ts": 1832,
     "apps/api/src/submissions.test.ts": 4804,
-    "apps/api/src/submissions.ts": 14600,
+    "apps/api/src/submissions.ts": 14589,
     "apps/api/src/suggestion-inbox.test.ts": 538,
     "apps/api/src/suggestion-inbox.ts": 791,
     "apps/api/src/suggestion-outcomes.test.ts": 154,
@@ -405,8 +405,7 @@
     "apps/web/src/gamePageApi.ts": 59,
     "apps/web/src/gamePlayer.bridge.test.ts": 334,
     "apps/web/src/gamePlayer.hiddenFrame.test.tsx": 97,
-    "apps/web/src/gamePlayer.test.ts": 44,
-    "apps/web/src/gamePlayer.ts": 1659,
+    "apps/web/src/gamePlayer.ts": 1630,
     "apps/web/src/gamePlayerBridge.test.ts": 528,
     "apps/web/src/gameSave.test.tsx": 281,
     "apps/web/src/gameSave.ts": 756,
@@ -506,7 +505,7 @@
     "apps/web/src/ScorecardPanel.test.tsx": 185,
     "apps/web/src/ScorecardPanel.tsx": 388,
     "apps/web/src/selfBuildCopy.test.ts": 102,
-    "apps/web/src/selfBuildCopy.ts": 459,
+    "apps/web/src/selfBuildCopy.ts": 420,
     "apps/web/src/sensing.test.tsx": 206,
     "apps/web/src/sensing.ts": 1025,
     "apps/web/src/ShareGameButton.test.tsx": 35,
@@ -543,7 +542,7 @@
     "apps/web/src/styles.studioShell.test.ts": 42,
     "apps/web/src/submissionApi.ts": 1268,
     "apps/web/src/SubmissionStatusView.test.ts": 1816,
-    "apps/web/src/SubmissionStatusView.tsx": 4141,
+    "apps/web/src/SubmissionStatusView.tsx": 4114,
     "apps/web/src/SuggestionsPanel.test.tsx": 28,
     "apps/web/src/SuggestionsPanel.tsx": 113,
     "apps/web/src/telemetry.test.ts": 201,
